refactor(scrape): replace github scraper with forge using ForgeReader trait
Phase 2 step 4: scrape/github → scrape/forge
- New scrape/forge/mod.rs uses ForgeReader trait for platform abstraction
- Detects forge from git remote URL, handles graceful degradation
- repo/internal.rs updated to use forge::run_legacy for backward compat
- Old scrape/github/mod.rs deleted (functionality moved to src/forge/)
- Event type changed from github.issue to forge.issue
