+++ src/commands/scrape/mod.rs
@@ -2,8 +2,8 @@
 
 pub mod code;
 pub mod database;
+pub mod forge;
 pub mod git;
-pub mod github;
 pub mod layer;
 pub mod sessions;
